leonardo_da_vqgan/models/clip/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.9.1+cu111
numpy
pillow
torchvision
ftfy
regex